requirements.txt
pandas
numpy
matplotlib
seaborn
scipy
networkx
plotly
biopython
sklearn
